added policy for comments, changed project and comemnt controllers and changed frontend

+++ app/Http/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Comment;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CommentController extends Controller
 {
@@ -17,19 +18,16 @@ class CommentController extends Controller
     public function store(Request $request, $project)
     {
         // dd($request);
+        $this->authorize('create', Comment::class);
         $data = Project::find($project);
 
         $request->validate([
             'content' => 'required|string|max:1000',
         ]);
 
-
-        // $user = auth()->user();
-
         $comment = new Comment();
         $comment->content = $request['content'];
-        $comment->user_id = $request['user_id'];
-        // $comment->user_id = $user->id;
+        $comment->user_id = Auth::id();
         $comment->project_id = $request['project_id'];
 
         $comment->save();
@@ -37,23 +35,13 @@ class CommentController extends Controller
         return view('comments.createComment', ['project' => $data]);
     }
 
-    // Retrieve a specific comment
-    // public function show(Project $project, Comment $comment)
-    // {
-    //     return view('comments.show', ['comment' => $comment]);
-    // }
-
     // Update an existing comment
     public function edit($project, $comment)
     {
+
         $project = Project::findOrFail($project);
         $comment = Comment::findOrFail($comment);
-
-        // Ensure the authenticated user is the owner of the comment
-        // if ($comment->user_id !== auth()->user()->id) {
-        //     return redirect()->route('projects.show', ['project' => $project->id])
-        //         ->withErrors('You do not have permission to edit this comment.');
-        // }
+        $this->authorize('update', $comment);
 
         return view('comments.editComment', ['project' => $project, 'comment' => $comment]);
     }
@@ -65,13 +53,7 @@ class CommentController extends Controller
 
         $project = Project::findOrFail($project);
         $comment = Comment::findOrFail($comment);
-        // dd($project, $comment);
-        // Ensure the authenticated user is the owner of the comment
-        // if ($comment->user_id !== auth()->user()->id) {
-        //     // If the user is not authorized, redirect them back
-        //     return redirect()->route('projects.show', ['project' => $project->id])
-        //         ->withErrors('You do not have permission to edit this comment.');
-        // }
+        $this->authorize('update', $comment);
 
         // Validate the input data
         $request->validate([
@@ -87,14 +69,10 @@ class CommentController extends Controller
     // // Delete a specific comment
     public function destroy($project, $comment)
     {
+
         $project = Project::findOrFail($project);
         $comment = Comment::findOrFail($comment);
-
-        // Ensure the authenticated user is the owner of the comment
-        // if ($comment->user_id !== auth()->user()->id) {
-        //     return redirect()->route('projects.show', ['project' => $project])
-        //         ->withErrors('You do not have permission to delete this comment.');
-        // }
+        $this->authorize('delete', $comment);
 
         $comment->delete();
